backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
cloudinary==1.39.0
pillow==10.2.0
python-dotenv==1.0.1
httpx==0.26.0